Fixed bug in lf. Erroneously used continue where I should have suppressed.

+++ src/include/cm.h
@@ -758,7 +758,7 @@ extern int answer_yn(char *, char *, char *, char *);
 extern int display_ok_message(char *);
 extern int Perror(char *);
 extern void user_end();
-extern bool lf_find(const char *, const char *, const char *, int, int);
+extern bool lf_find(const char *, const char *, const char *, int, long);
 extern size_t canonicalize_file_spec(char *);
 extern bool construct_file_spec(char *, char *, char *, char *, char *, int);
 extern bool file_spec_path(char *, char *);
